重构: 将 utf8/gbk 互转的代码移动到 libbmy/convcode
